requirements.txt
django>=3.1
pymysql
djangorestframework
pygments